Rework WP source files and update documentation
- rearrange location of several Wire Protocol source files
- update documentation on Wire Protocol
- minor updates on other documentation files
