+++ targets/CMSIS-OS/ChibiOS/ST_NUCLEO_F091RC/README.md
@@ -1,10 +1,13 @@
-For the ChibiOS test app to build succesfully the following changes are required:
+For the a successful build the following changes are required:
 
-In _halconf.g_ (when compared with a default file)
+In _halconf.g_ (in both nanoBooter and nanoCLR folders), when compared with a default file:
 - HAL_USE_SERIAL to TRUE
 - SERIAL_DEFAULT_BITRATE to 115200
 
-In _mcuconf.h_ (when compared with a default file)
+In _mcuconf.h_ (in both nanoBooter and nanoCLR folders), when compared with a default file:
 - STM32_SERIAL_USE_USART2 to TRUE
 
-NOTE: this configuration was sucessfully tested in a NUCLEO_F091RC board using the virtual COM port through the ST Link USB connection.
+In _chconf.h_ (_**only**_ for nanoCLR folder), when compared with a default file:
+- set the CORTEX_VTOR_INIT with the appropriate address of the vector table for nanoCLR
+
+NOTE: this configuration was successfully tested in a NUCLEO_F091RC board using the virtual COM port through the ST Link USB connection.
