+++ targets/CMSIS-OS/ChibiOS/nanoBooter/WireProtocol_App_Interface.c
@@ -1,5 +1,6 @@
 //
 // Copyright (c) 2017 The nano Framework project contributors
+// Portions Copyright (c) Microsoft Corporation.  All rights reserved.
 // See LICENSE file in the project root for full license information.
 //
