+++ targets/CMSIS-OS/ChibiOS/nanoBooter/WireProtocol_MonitorCommands.c
@@ -4,7 +4,8 @@
 // See LICENSE file in the project root for full license information.
 //
 
-#include "WireProtocol_Commands.h"
+#include <WireProtocol.h>
+#include <WireProtocol_MonitorCommands.h>
 
 //////////////////////////////////////////////////////////////////////
 // helper functions
@@ -17,10 +18,11 @@ bool NanoBooter_GetReleaseInfo(ReleaseInfo* releaseInfo)
     releaseInfo->version.usRevision = 1;
 
     // TODO replace this with string from (possibly...) main config file
-    memcpy(&releaseInfo->infoString, ">>>> nanoFramework RULES <<<<", sizeof(releaseInfo->infoString));
+    memcpy(&releaseInfo->infoString, ">>>> nanoFramework RULES (nanoBooter here) <<<<", sizeof(releaseInfo->infoString));
 
     return true;
 }
+
 
 ////////////////////////////////////////////////////
 
@@ -29,7 +31,7 @@ bool Monitor_Ping(WP_Message* message)
     if((message->m_header.m_flags & WP_Flags_c_Reply) == 0)
     {
         Monitor_Ping_Reply cmdReply;
-        cmdReply.m_source = Monitor_Ping_c_Ping_Source_NanoCLR;
+        cmdReply.m_source = Monitor_Ping_c_Ping_Source_NanoBooter;
 
         ReplyToCommand(message, true, false, &cmdReply, sizeof(cmdReply));
     }
@@ -39,13 +41,13 @@ bool Monitor_Ping(WP_Message* message)
 
 bool Monitor_OemInfo(WP_Message* message)
 {
-    if((message->m_header.m_flags & WP_Flags_c_Reply      ) == 0)
+    if((message->m_header.m_flags & WP_Flags_c_Reply) == 0)
     {
         Monitor_OemInfo_Reply cmdReply;     
         
         bool fOK = NanoBooter_GetReleaseInfo(&cmdReply.m_releaseInfo) == true;
         
-        ReplyToCommand(message, fOK, false, &cmdReply, sizeof(cmdReply) );
+        ReplyToCommand(message, fOK, false, &cmdReply, sizeof(cmdReply));
     }
 
     return true;
